Refactor HeaderMenu to new directory
